Searches Done
This is kind of the final version of my parts of the project.
With this update, the user can use the anatomy of a body to select where they feel the discomfort, then they can choose up to four symptoms of their illness. When they press search, a new window with their diagnosis opens, telling them a brief summary of the illness they most likely have, recommended medecin, recommended doctor, and a few other possible illnesses. There is also another type of search, which takes the input from the user (3 symptoms and the general area of discomfort) and opens it in a google search as a new JFrame. From there, they can read about their most likely diagnosis and find their needed medecin to purchase on the app.

This version does not handle error and mostly assumin the user enters what is asked of them to enter. I will now start debugging and fool-proofing it!